// frontend/src/api/model/CreateScheduleCommand.ts
/**
 * Generated by orval v8.7.0 🍺
 * Do not edit manually.
 * TeleHealth.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { LocalDate } from "./LocalDate";
import type { LocalTime } from "./LocalTime";

export interface CreateScheduleCommand {
  doctorPublicId: string;
  date: LocalDate;
  startTime: LocalTime;
  endTime: LocalTime;
  scheduleStatus: string;
}